feat: Initalise terrafom project
Introduce the terraform setup for the production homelab. The intention
at the moment is to keep things very simple, with Kubernetes/Clustering
done at a later date as an experimental lab.

"Production" services are things that are self-hosted, that I can't live
without (password manager, media servers, kanboard, etc.)

Secrets are SOPs encrypted. the Ansible inventory is generated by
Terraform.