feat: Add ansible playbooks
These playbooks setup the servers provisioned with terraform. The goal
is to keep things simple and maintainable so, even if swamped with work,
I can still "keep the lights on".

It consists of a playbook for each Linux Host. And a role to setup
Samba. Due to deficiencies in the Proxmox API (and therefore terraform
provider) we need to perform additional configuration on by manually
editing container confituration files on the Proxmox host (chireiden).

Secrets are handled with SOPS.