feat(hashihime): Add jellyfin
Jellyfin was setup on the LXC container hashihime. A role for installing
Nvidia drivers on the host and container was created. Installation of
jellyfin was also implemented as a role.

For the reverse proxy satori some changes to the https_simple_site role
were made to facilitate custom directives in the server block.

Some refactoring of the production group vars file for services was also
performed.